Thermostat Firmware Update
Rewriting the code. Taking information from the original code and
supplying it to an updated version.

Updates include:
- Rewritten Board_Interface class: Taken from old BMPSense class. The
Board_Interface is now created using pointers and a singleton design.
- Rewritten States class and subclasses: The old states weren't working
properly from another branch. Now the new states will utilize the
Board_Interface object to help with consistency. Duties are also going
to be more explicitly separated, and better handled by each state.
- Main .ino program file: It will now utilize the state machine setup,
with the singleton object for the sensor and SD card.